Added Stripe Payment (removed secret key)

+++ src/main/java/com/airBnbClone/AirBnbClone/service/HotelServiceImpl.java
@@ -5,13 +5,16 @@ import com.airBnbClone.AirBnbClone.Dto.HotelInfoDto;
 import com.airBnbClone.AirBnbClone.Dto.RoomDto;
 import com.airBnbClone.AirBnbClone.entity.Hotel;
 import com.airBnbClone.AirBnbClone.entity.Room;
+import com.airBnbClone.AirBnbClone.entity.User;
 import com.airBnbClone.AirBnbClone.exception.ResourceNotFoundException;
+import com.airBnbClone.AirBnbClone.exception.UnAuthorizedException;
 import com.airBnbClone.AirBnbClone.repository.HotelRepository;
 import com.airBnbClone.AirBnbClone.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,15 +34,22 @@ public class HotelServiceImpl implements HotelService {
         log.info("createNewHotel");
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
-        Hotel saveHotel = hotelRepository.save(hotel);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+        hotel = hotelRepository.save(hotel);
         log.info("Hotel created successfully");
-        return modelMapper.map(saveHotel , HotelDto.class);
+        return modelMapper.map(hotel , HotelDto.class);
     }
 
     @Override
     public HotelDto getHotelById(Long id) {
         log.info("Getting the hotel with ID: {}", id);
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Not found"));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+id);
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -47,9 +57,13 @@ public class HotelServiceImpl implements HotelService {
     public HotelDto updateHotelById(Long id, HotelDto hotelDto) {
         log.info("Updating the hotel with ID: {}", id);
         Hotel existing = hotelRepository.findById(id).orElseThrow(() ->new  ResourceNotFoundException("Not found"));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(existing.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+id);
+        }
         modelMapper.map(hotelDto , existing);
         existing.setId(id);
-        hotelRepository.save(existing);
+        existing = hotelRepository.save(existing);
         return modelMapper.map(existing, HotelDto.class);
     }
 
@@ -63,18 +77,29 @@ public class HotelServiceImpl implements HotelService {
     @Transactional
     public void deleteHotelById(Long id) {
         Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Not found"));
-        hotelRepository.deleteById(id);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+id);
+        }
         for(Room room : hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
         }
         log.info("Hotel deleted successfully");
+        hotelRepository.deleteById(id);
     }
 
     @Override
+    @Transactional
     public void activeHotel(Long hotelId) {
         log.info("Activating the hotel with ID: {}", hotelId);
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Not found"));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         hotel.setActive(true);
 
         for(Room room : hotel.getRooms()){
