+++ src/main/java/com/airBnbClone/AirBnbClone/service/RoomServiceImpl.java
@@ -3,13 +3,16 @@ package com.airBnbClone.AirBnbClone.service;
 import com.airBnbClone.AirBnbClone.Dto.RoomDto;
 import com.airBnbClone.AirBnbClone.entity.Hotel;
 import com.airBnbClone.AirBnbClone.entity.Room;
+import com.airBnbClone.AirBnbClone.entity.User;
 import com.airBnbClone.AirBnbClone.exception.ResourceNotFoundException;
+import com.airBnbClone.AirBnbClone.exception.UnAuthorizedException;
 import com.airBnbClone.AirBnbClone.repository.HotelRepository;
 import com.airBnbClone.AirBnbClone.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,23 +31,32 @@ public class RoomServiceImpl implements RoomService {
     public RoomDto createNewRoom(Long hotelId, RoomDto roomDto) {
         log.info("Room creation started");
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Not found"));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+hotelId);
+        }
         log.info("hotel id found");
         Room room = modelMapper.map(roomDto, Room.class);
         log.info("room id found and mapped to dto");
         room.setHotel(hotel);
         log.info("Set to hotel");
-        Room saveRoom = roomRepository.save(room);
+        room = roomRepository.save(room);
         log.info("Room created");
         if (hotel.getActive()) {
             inventoryService.initializeRoomForAYear(room);
         }
-        return modelMapper.map(saveRoom, RoomDto.class);
+        return modelMapper.map(room, RoomDto.class);
     }
 
     @Override
     public List<RoomDto> getAllRoomsInHotel(Long hotelId) {
         log.info("Getting all rooms in hotel with ID: {}", hotelId);
         Hotel hotel = hotelRepository.findById(hotelId).orElseThrow(() -> new ResourceNotFoundException("Not found"));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorizedException("This user does not own this hotel with id: "+hotelId);
+        }
         return hotel.getRooms()
                 .stream()
                 .map(element -> modelMapper.map(element, RoomDto.class))
@@ -59,12 +71,6 @@ public class RoomServiceImpl implements RoomService {
         return modelMapper.map(room, RoomDto.class);
     }
 
-    public void isExistHotel(Long hotelId) {
-        boolean exist = hotelRepository.existsById(hotelId);
-        if (!exist) {
-            throw new ResourceNotFoundException("Not found");
-        }
-    }
 
     public void isExistRoom(Long roomId) {
         boolean exist = roomRepository.existsById(roomId);
@@ -77,9 +83,15 @@ public class RoomServiceImpl implements RoomService {
     @Override
     public void deleteRoomById(Long roomId) {
         log.info("Deleting the room with ID: {}", roomId);
+
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: " + roomId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorizedException("This user does not own this room with id: "+roomId);
+        }
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
         log.info("Room deleted");
